backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
tweepy==4.14.0
praw==7.7.1
vaderSentiment==3.3.2
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
gunicorn==21.2.0
ratelimit==2.2.1 